feat: 검증된 Participants 모델 기능 구현
