+++ README.md
@@ -47,7 +47,7 @@
 - 참여자
   - [x] 참여자는 딜러 혹은 플레이어가 된다.
 - 참여자들
-  - [ ] 참여자들은 참여자 목록을 가지고 있다.
+  - [x] 참여자들은 참여자 목록을 가지고 있다.
 ---
 - Manager
   - 게임 매니저는 게임 규칙을 관리한다.
